Fix to menus, so that you can change the anchor URL and it will respond.
No longer do anything in onclick listeners for <a> tags in the navbar, but
rather wait for the browser to send us a notification that the anchor tag
has called.

Also included bootstrap.min.js, so the collapsed menu works in responsive views as well.

Tweak to menu highlighting to remove duplicate code.