Made the setup/setup-db.sh script work without input.
Now it reads a bunch of scripts and executes them with sqlite3.
Previously, it output sql to be executed within sqlite (what was
I thinking?).

It also includes some stuff to do with calculating relative rankings
of attributes for each LGA, to be used in the future by the card
game which will be merged in the future.