No longer require absolute URL's.
These were from when we were developing in a local .html file,
with no live server. When we added a server, it was on a different
host, and so we needed the absolute URL. Now though, it is relatively
straightforward to get the server running locally.

The only place PB.env was being used was to differentiate between
URLs to use. By removing this, it now also doesn't matter what host
the server binds to, and it also doesn't matter which port either.

If in the future there is a need to re-add a distinction between dev
and live environments, we can do it then. But until then, may as well
leave it out as it was cluttering the git staging area (leaving it
in dev mode locally but commiting with live mode).